Modulized Space Weather tool, moved output directores of scripts to noodle
